Finished single-sided-scan.py and testing it.
Appears to work. Also cleaned up a bunch of messages and displaying.
I decided to output only minimal things to /home/arjun/brscan/brscan.log
and oinly noting the script name there. The rest is stored in the python
logfile for the indivudual scripts.

To do: remove references to your own directories, and instead make some
generic directory where the logs are stored. This is so that others can
use it.