Resume at integrating the double-sided scanning script.
Needs a is_duplex_capable function which simply scans the --source line
from scanimage -h, sees if the scanner is duplex capable, and if the
source contains a duplex argument. If the source does not contain a
duplex argument, simply run the single sided scan routine. I should read
the argument that says --double. Anyway, this requires a bit of work,
should probably draw a decision tree.