Debugged scantoocr in auto mode on thinkpad.
It has to do with the location of the configuration file. I had to use
readlink to make sure I got the actual directory. This is because in
brscan's default setting, it uses

    /opt/brother/scanner/brscan-skey/brscan-skey-*.cfg

for configuration and the script is in

    /opt/brother/scanner/brscan-skey/script

So the config file should be referenced as

    ../brscan-skey-*.cfg

so that it's a bit more robust.